tests/integration/test_full_workflow.py
# Copyright RealTerrain Studio. All Rights Reserved.

"""
Integration tests for complete RealTerrain Studio workflow.
"""

import pytest
import json
import zipfile
from pathlib import Path
from PIL import Image
import numpy as np


class TestCompleteWorkflow:
    """Test complete terrain export to UE5 import workflow."""

    def test_export_minimal_terrain(self, temp_dir):
        """Test exporting a minimal terrain package."""
        # This would test the complete export from QGIS
        # For now, verify structure manually

        # Expected output structure
        expected_files = [
            'heightmap.png',
            'metadata.json',
            'satellite_texture.png',
            'osm_splines.json'
        ]

        # Test that we can create these files
        for filename in expected_files:
            filepath = temp_dir / filename
            assert isinstance(filepath, Path)

    def test_metadata_completeness(self, temp_dir, sample_metadata):
        """Test that metadata contains all required fields."""
        metadata_path = temp_dir / "metadata.json"

        # Write metadata
        with open(metadata_path, 'w') as f:
            json.dump(sample_metadata, f, indent=2)

        # Read and verify
        with open(metadata_path) as f:
            data = json.load(f)

        # Verify required fields
        assert 'heightmap' in data
        assert 'export_info' in data

        # Verify heightmap metadata
        heightmap = data['heightmap']
        assert 'width' in heightmap
        assert 'height' in heightmap
        assert 'min_elevation' in heightmap
        assert 'max_elevation' in heightmap
        assert 'pixel_size_x' in heightmap
        assert 'pixel_size_y' in heightmap
        assert 'crs' in heightmap

    def test_heightmap_satellite_alignment(self, temp_dir):
        """Test that heightmap and satellite texture are properly aligned."""
        # Create test heightmap
        heightmap_size = (1024, 1024)
        heightmap_data = np.random.randint(0, 65535, heightmap_size, dtype=np.uint16)
        heightmap_path = temp_dir / "heightmap.png"

        # Create test satellite texture (typically 2x resolution)
        satellite_size = (2048, 2048)
        satellite_data = np.random.randint(0, 255, (*satellite_size, 3), dtype=np.uint8)
        satellite_path = temp_dir / "satellite_texture.png"

        # In actual implementation, verify they cover the same geographic area
        # For now, just verify size relationship
        assert satellite_size[0] >= heightmap_size[0]
        assert satellite_size[1] >= heightmap_size[1]

    def test_osm_splines_structure(self, temp_dir):
        """Test OSM splines JSON structure."""
        osm_data = {
            'roads': [
                {
                    'spline_id': 'way_12345',
                    'name': 'Main Street',
                    'road_type': 'primary',
                    'width': 800,
                    'lanes': 2,
                    'surface': 'asphalt',
                    'one_way': False,
                    'points': [
                        {'position': [100.0, 200.0, 10.0]},
                        {'position': [150.0, 250.0, 12.0]}
                    ]
                }
            ],
            'railways': [],
            'power_lines': []
        }

        osm_path = temp_dir / "osm_splines.json"
        with open(osm_path, 'w') as f:
            json.dump(osm_data, f, indent=2)

        # Verify structure
        with open(osm_path) as f:
            data = json.load(f)

        assert 'roads' in data
        assert 'railways' in data
        assert 'power_lines' in data
        assert len(data['roads']) == 1
        assert data['roads'][0]['spline_id'] == 'way_12345'


class TestDataValidation:
    """Test data validation across the pipeline."""

    def test_heightmap_data_range(self, mock_heightmap_data):
        """Test that heightmap data is in valid range."""
        assert mock_heightmap_data.dtype == np.uint16
        assert np.min(mock_heightmap_data) >= 0
        assert np.max(mock_heightmap_data) <= 65535

    def test_coordinate_system_consistency(self, sample_metadata):
        """Test that coordinate system is consistent."""
        crs = sample_metadata['heightmap']['crs']
        assert crs in ['EPSG:4326', 'EPSG:3857', 'EPSG:32633']  # Common CRS

    def test_elevation_range_valid(self, sample_metadata):
        """Test that elevation range is valid."""
        min_elev = sample_metadata['heightmap']['min_elevation']
        max_elev = sample_metadata['heightmap']['max_elevation']

        assert max_elev > min_elev
        assert min_elev >= -500.0  # Dead Sea is around -430m
        assert max_elev <= 9000.0  # Everest is 8849m


class TestErrorRecovery:
    """Test error recovery and graceful degradation."""

    def test_missing_satellite_texture(self, temp_dir):
        """Test that import works without satellite texture."""
        # Should be able to import with just heightmap and metadata
        # Satellite texture is optional
        assert True  # Placeholder

    def test_missing_osm_splines(self, temp_dir):
        """Test that import works without OSM splines."""
        # Should be able to import with just heightmap
        # OSM splines are optional
        assert True  # Placeholder

    def test_partial_osm_data(self, temp_dir):
        """Test handling of partial OSM data."""
        # Should handle cases where only roads exist, no railways/powerlines
        osm_data = {
            'roads': [
                {
                    'spline_id': 'way_12345',
                    'name': 'Test Road',
                    'road_type': 'residential',
                    'width': 600,
                    'lanes': 1,
                    'surface': 'asphalt',
                    'one_way': False,
                    'points': [
                        {'position': [0.0, 0.0, 0.0]},
                        {'position': [100.0, 0.0, 0.0]}
                    ]
                }
            ],
            'railways': [],  # Empty
            'power_lines': []  # Empty
        }

        osm_path = temp_dir / "osm_splines.json"
        with open(osm_path, 'w') as f:
            json.dump(osm_data, f)

        # Should handle gracefully
        assert osm_path.exists()

    def test_corrupted_json(self, temp_dir):
        """Test handling of corrupted JSON files."""
        corrupted_path = temp_dir / "corrupted.json"
        with open(corrupted_path, 'w') as f:
            f.write("{invalid json content")

        # Should fail gracefully without crashing
        # In actual implementation, would test the parser
        assert corrupted_path.exists()


class TestPerformance:
    """Test performance of complete workflow."""

    @pytest.mark.performance
    def test_small_terrain_performance(self, temp_dir):
        """Test export performance for small terrain (1km²)."""
        # Small terrain should export very quickly (< 5 seconds)
        assert True  # Placeholder for actual timing test

    @pytest.mark.performance
    def test_medium_terrain_performance(self, temp_dir):
        """Test export performance for medium terrain (10km²)."""
        # Medium terrain should export in reasonable time (< 30 seconds)
        assert True  # Placeholder

    @pytest.mark.performance
    def test_large_terrain_performance(self, temp_dir):
        """Test export performance for large terrain (100km²)."""
        # Large terrain should complete (< 2 minutes)
        assert True  # Placeholder


class TestCrossCompatibility:
    """Test compatibility across different versions and platforms."""

    def test_ue5_compatibility(self, temp_dir):
        """Test that exported data is compatible with UE5."""
        # Verify data formats match UE5 expectations
        assert True  # Placeholder

    def test_coordinate_system_support(self, temp_dir):
        """Test support for multiple coordinate systems."""
        # Should support EPSG:4326 (WGS84), EPSG:3857 (Web Mercator), etc.
        supported_crs = ['EPSG:4326', 'EPSG:3857']
        assert len(supported_crs) > 0

    def test_platform_independence(self, temp_dir):
        """Test that exports work on Windows, Mac, Linux."""
        # Path handling should be platform-independent
        test_path = Path(temp_dir) / "test.png"
        assert isinstance(test_path, Path)


# Smoke tests - quick verification that basic functionality works
@pytest.mark.smoke
class TestSmokeTests:
    """Quick smoke tests for basic functionality."""

    def test_can_create_heightmap(self, temp_dir, mock_heightmap_data):
        """Smoke test: Can create a heightmap file."""
        heightmap_path = temp_dir / "heightmap.png"
        # In actual implementation: save mock_heightmap_data as PNG
        assert mock_heightmap_data.shape == (64, 64)

    def test_can_create_metadata(self, temp_dir, sample_metadata):
        """Smoke test: Can create metadata file."""
        metadata_path = temp_dir / "metadata.json"
        with open(metadata_path, 'w') as f:
            json.dump(sample_metadata, f)
        assert metadata_path.exists()

    def test_can_parse_json(self, temp_dir):
        """Smoke test: Can parse JSON files."""
        test_data = {'test': 'value'}
        test_path = temp_dir / "test.json"
        with open(test_path, 'w') as f:
            json.dump(test_data, f)

        with open(test_path) as f:
            loaded = json.load(f)

        assert loaded['test'] == 'value'
